+++ CanHazFunny/CanHazFunny/JokeOutput.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Net.Http;
 
 namespace CanHazFunny;
@@ -10,6 +11,8 @@ public class JokeOutput : IOutput
         ArgumentNullException.ThrowIfNull(joke);
         Console.WriteLine(joke);
     }
+
+   
 
 }
